linux: fix to comply with XDG, and show correct name in GNOME Trash
> When trashing a file or directory, the implementation MUST create the
> corresponding file in $trash/info first. Moreover, it MUST try to do this
> in an atomic fashion, so that if two processes try to trash files with
> the same filename this will result in two different trash files. On
> Unix-line systems this is done by generating a filename, and then opening
> with O_EXCL. If that succeeds the creation was atomic (at least on the
> same machine), if it fails you need to pick another filename.
>
> https://specifications.freedesktop.org/trash-spec/trashspec-latest.html

Except, regarding the last sentence, if our UUIDs are not unique then
something is very wrong.  In that case, let's just fail with 'EEXIST'.

It looks like GnomeVFS was sanity-checking the timestamps.  So if you break
the rule and don't create the trashinfo file first, then GnomeVFS will
ignore it :-).

Fixes issue #56